# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer DB API",
        "description": "Customer database API: authentication with access and refresh tokens, customer management, avatars, CSV import/export and background jobs.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "v1.0.0"
    },
    "paths": {
...
        "/customer": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "List customers",
                "description": "Paginated customer list with keyword search and status filter.",
                "operationId": "39c5716604f7dccc3f995c28240dda86",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "keywords",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "lead, prospect, active or inactive",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Create customer",
                "description": "Create a customer. Requires an admin token.",
                "operationId": "a882d2d5f5e63ea16a169132edb19542",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "lead",
                                            "prospect",
                                            "active",
                                            "inactive"
                                        ]
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    },
                    "403": {
                        "description": "Access denied"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            }
        },
        "/customer/stats": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Customer statistics",
                "description": "Customer counts per status, total and creations in the last seven days.",
                "operationId": "29eec06d801b4095ee38e6bd727c918c",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            }
        },
        "/customer/{id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Customer detail",
                "description": "Retrieve a single customer.",
                "operationId": "f9e92274ac283706b84ba139e36580de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Update customer",
                "description": "Update a customer. Requires an admin token.",
                "operationId": "29e4c37f8669b25de09c1f37cbbbae62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "lead",
                                            "prospect",
                                            "active",
                                            "inactive"
                                        ]
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer updated"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Delete customer",
                "description": "Soft-delete a customer and its avatar. Requires an admin token.",
                "operationId": "dd5c98734b71e95c10efe43fe7e0b2f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer deleted"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            }
        },
        "/customer/{id}/avatar": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Upload avatar",
                "description": "Upload or replace the customer avatar (JPEG, PNG or WebP). Requires an admin token.",
                "operationId": "fe334c23f74562ac8f7bbf7ec975f710",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar stored"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Remove avatar",
                "description": "Remove the customer avatar. Requires an admin token.",
                "operationId": "3f52c6ef7e7adfd11fa5de105d00cff3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avatar removed"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "auth_token": {
                "type": "apiKey",
                "description": "JWT issued by the application; send the raw token or 'Bearer <token>'.",
                "name": "Authorization",
                "in": "header"
            },
            "health_token": {
                "type": "apiKey",
                "description": "Token configured through HEALTHCHECK_TOKEN for detailed health diagnostics.",
                "name": "X-Health-Token",
                "in": "header"
            }
        }
    },
...
}